+++ src/promptflow/promptflow/_sdk/_submitter/run_submitter.py
@@ -8,22 +8,22 @@ from pathlib import Path
 from typing import Union
 
 from promptflow._constants import FlowLanguage
-from promptflow._core.operation_context import OperationContext
 from promptflow._sdk._constants import ContextAttributeKey, FlowRunProperties
-from promptflow._sdk._utils import parse_variant
 from promptflow._sdk.entities._flow import Flow
 from promptflow._sdk.entities._run import Run
 from promptflow._sdk.operations._local_storage_operations import LocalStorageOperations
 from promptflow._utils.context_utils import _change_working_dir
+from promptflow._utils.flow_utils import parse_variant
 from promptflow.batch import BatchEngine
 from promptflow.contracts.run_info import Status
 from promptflow.contracts.run_mode import RunMode
 from promptflow.exceptions import UserErrorException, ValidationException
+from promptflow.tracing._operation_context import OperationContext
 
 from ..._utils.logger_utils import LoggerFactory
 from .._configuration import Configuration
 from .._load_functions import load_flow
-from ..entities._eager_flow import FlexFlow
+from ..entities._flow import FlexFlow
 from .utils import SubmitterHelper, variant_overwrite_context
 
 logger = LoggerFactory.get_logger(name=__name__)
@@ -105,16 +105,17 @@ class RunSubmitter:
     def _submit_bulk_run(self, flow: Union[Flow, FlexFlow], run: Run, local_storage: LocalStorageOperations) -> dict:
         logger.info(f"Submitting run {run.name}, log path: {local_storage.logger.file_path}")
         run_id = run.name
-        if flow.language == FlowLanguage.CSharp:
-            # TODO: consider moving this to Operations
-            from promptflow.batch import CSharpExecutorProxy
+        # for python, we can get metadata in-memory, so no need to dump them first
+        if flow.language != FlowLanguage.Python:
+            from promptflow._proxy import ProxyFactory
 
-            CSharpExecutorProxy.generate_metadata(flow_file=Path(flow.path), assembly_folder=Path(flow.code))
-            # TODO: shall we resolve connections here?
-            connections = []
-        else:
-            with _change_working_dir(flow.code):
-                connections = SubmitterHelper.resolve_connections(flow=flow)
+            # variants are resolved in the context, so we can't move this logic to Operations for now
+            ProxyFactory().get_executor_proxy_cls(flow.language).dump_metadata(
+                flow_file=Path(flow.path), working_dir=Path(flow.code)
+            )
+
+        with _change_working_dir(flow.code):
+            connections = SubmitterHelper.resolve_connections(flow=flow)
         column_mapping = run.column_mapping
         # resolve environment variables
         run.environment_variables = SubmitterHelper.load_and_resolve_environment_variables(