+++ src/promptflow/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.7.0 (Upcoming)
 
+### NOTICES
+- Import warnings will be printed when importing from `promptflow` namespace, please use imports from new namespaces
+  suggested in the warning message.
+
 ### Features Added
 
 - [SDK/CLI] Create a run with `resume_from`, note that only run created with `promptflow>=1.7.0` can be used as the value of `resume_from`:
@@ -11,12 +15,15 @@
   - CLI: Support `pfazure run create --resume-from <original-run-name>` to create a run resume from another run.
   - SDK: Support `pf.run(resume_from=<original-run-name>)` to create a run resume from another run.
 
+- [Batch] Added per-line logging for batch runs, stored under the `flow_logs` folder.
 - [SDK/CLI] Support `AzureOpenAIConnection.from_env` and `OpenAIConnection.from_env`. Reach more details [here](https://microsoft.github.io/promptflow/how-to-guides/manage-connections.html#load-from-environment-variables).
 
 ### Bugs Fixed
 
 - [SDK/CLI] environment variable `PF_HOME_DIRECTORY` doesn't work for run details & logs.
+- [SDK/CLI] Support override hard coded "deployment_name" and "model".
 - [SDK] `connection.provider` config doesn't work when calling flow as a function.
+- [SDK/CLI] Support override unprovided connection inputs in nodes.
 
 ## 1.6.0 (2024.03.01)